+++ apps/heft/src/operations/runners/PhaseOperationRunner.ts
@@ -57,8 +57,10 @@ export class PhaseOperationRunner implements IOperationRunner {
     }
 
     // Delete the files if any were specified
-    const rootFolderPath: string = internalHeftSession.heftConfiguration.buildFolderPath;
-    await deleteFilesAsync(rootFolderPath, deleteOperations, cleanLogger.terminal);
+    if (deleteOperations.length > 0) {
+      const rootFolderPath: string = internalHeftSession.heftConfiguration.buildFolderPath;
+      await deleteFilesAsync(rootFolderPath, deleteOperations, cleanLogger.terminal);
+    }
 
     // Ensure we only run the clean operation once
     this._isClean = true;
